[FEAT] Split up offer components
